# David Osideko- HSBC/openapi.v1.yaml
openapi: 3.0.0
info:
  title: HSBC Library API
  version: '1.0'
  description: Interface for managing book collections and rentals
  contact:
    name: David Osideko
    email: david.osideko@hsbc.co.uk
servers:
  - url: 'http://localhost:8000/v1'
    description: Local API server
  - url: 'https://hsbc.co.uk/library/api/v1'
    description: Live API server
paths:
...
  /member:
    post:
      summary: Member creation
      operationId: createMember
      responses:
        '201':
          description: User successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              examples: {}
        '400':
          description: A required field was not provided
        '409':
          description: The provided email already exists
        '500':
          description: An error occurred while creating the member record
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInfo'
            examples:
              Create User Matt:
                value:
                  firstName: Matt
                  lastName: Damon
                  email: matt.damon@hsbc.co.uk
                  gender: MALE
        description: Provide the necessary fields for user creation
      description: Create a new user
      tags:
        - Membership
    parameters: []
  '/member/{id}':
    parameters:
      - schema:
          type: string
        name: id
        in: path
        required: true
        description: The member's user identifier
    get:
      summary: Member retrieval
      tags:
        - Membership
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              examples: {}
        '404':
          description: A user with the provided identifier was not found
        '500':
          description: Error while retrieving the user's details
      operationId: getMemberById
      description: Retrieve a single library member associated with the provided unique identifier
    put:
      summary: Member update
      operationId: updateMemberDetails
      responses:
        '200':
          description: OK - Details successfully updated
        '400':
          description: Bad Request - invalid input data provided
        '500':
          description: An error occurred while updating this record
      description: Update the details of the user with the corresponding id
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInfo'
            examples:
              Update Matt's Profile:
                value:
                  firstName: Matt
                  lastName: Hancock
                  email: matt.hancock@hsbc.co.uk
                  gender: FEMALE
                  isStaff: true
                  isLibrarian: true
        description: Provide any user detail which you intend to update
      tags:
        - Membership
    delete:
      summary: Member deletion
      operationId: deleteMember
      responses:
        '200':
          description: OK - User successfully deleted
        '404':
          description: A user with the provided id does not exist
        '500':
          description: An error occurred while updating this record
      description: Delete the record of the user with the corresponding id
      tags:
        - Membership
...
  /catalogue/book:
    post:
      summary: Catalogue book creation
      operationId: addBook
      responses:
        '201':
          description: Detail created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '400':
          description: Bad Request - one or more of the provided inputs is invalid
        '500':
          description: An error occurred while adding this book to the catalogue
      description: Add a book to the catalogue
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookInfo'
            examples:
              The Power of Now:
                value:
                  author: Eckhart Tolle
                  title: The Power of Now
                  description: A Guide to Spiritual Enlightenment
                  isbn: '9780340733509'
                  publishedDate: 0
                  language: ENGLISH
                  genre: SELF_HELP
                  classification: B
                  copiesAvailable: 1
        description: Create a book record in the catalogue
      tags:
        - Catalogue
  '/catalogue/book/{uuid}':
    parameters:
      - schema:
          type: string
        name: uuid
        in: path
        required: true
        description: Universally unique identifier
    get:
      summary: Catalogue book retrieval
      tags:
        - Catalogue
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '400':
          description: An invalid UUID was provided
        '404':
          description: A book with the provided UUID was not found
        '500':
          description: An error occured while retrieving the book with the provided identifier
      operationId: getBookByUuid
      description: Retrieve a book by its universally unique identifier
    put:
      summary: Catalogue book details update
      operationId: updateBookDetails
      responses:
        '200':
          description: OK - Update successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book'
        '400':
          description: Bad Request - an invalid input was provided in the request
        '500':
          description: An error occurred while updating the book details
      description: Update the details of the book with the associated universally unique identifier
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookInfo'
            examples:
              The Power of Now:
                value:
                  author: Eckhart Tolle
                  title: The Power of Now
                  description: A Guide to Spiritual Enlightenment
                  isbn: '9780340733509'
                  publicationDate: '0'
                  language: ENGLISH
                  coverImageData: ''
                  genre: SELF_HELP
                  classification: B
                  copiesAvailable: 1
        description: Details of the book which require updating
      tags:
        - Catalogue
    delete:
      summary: Catalogue book deletion
      operationId: deleteBookByUuid
      responses:
        '200':
          description: Successfully deleted
        '404':
          description: A book with the provided identifier was not found
        '500':
          description: An error occurred while deleting the associated book
      description: Delete the book with the associated unique identifier
      tags:
        - Catalogue
  /rental:
    post:
      summary: Book rental creation
      operationId: createRental
      responses:
        '201':
          description: Book loan registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Loan'
        '404':
          description: Either a book with the provided uuid or member with the provided id was not found
        '500':
          description: An error occurred while creating the book loan record
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoanInfo'
            examples:
              Minimum loan requirement:
                value:
                  bookUuid: c5043628-d7d5-4969-b6b9-9ad8bb2a57b7
                  memberId: 1
        description: Provide details of the loan
      description: Create a record of a borrowed book
      tags:
        - Rental
  '/rental/{id}':
    parameters:
      - schema:
          type: string
        name: id
        in: path
        required: true
        description: Loan identifier
    get:
      summary: Book rental detail retrieval
      tags:
        - Rental
      responses:
        '200':
          description: OK - Successful retrieval
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Loan'
        '404':
          description: A rental record with the provided id does not exist
        '500':
          description: An error occured while retrieving the rental information
      operationId: getRentalId
      description: Retrieve the details of a single rental
    put:
      summary: Book rental detail update
      operationId: updateRentalDetail
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Loan'
      description: Update details relating to this book rental
      tags:
        - Rental
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoanInfo'
  '/library/{id}':
    parameters:
      - schema:
          type: string
        name: id
        in: path
        required: true
        description: Library identifier
    get:
      summary: Library details retrieval
      tags:
        - Library
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Library'
        '404':
          description: The library with the specified identifier was not found
        '500':
          description: An error occurred while retrieving library details
      operationId: getLibraryById
      description: Retrieve details of the library with the associated identifier
    put:
      summary: Library details update
      operationId: updateLibrary
      responses:
        '200':
          description: Library details updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Library'
        '400':
          description: Bad Request - invalid data provided in the payload
        '500':
          description: An error occured while updating library information
      description: Update the details of the library with the associated identifier
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LibraryInfo'
            examples:
              Test Canterbury Library:
                value:
                  name: Canterbury Library
                  address: Somewhere else in the UK
                  email: library@canterbury.co.uk
                  phoneNumber: '0'
        description: Provide library details to override
      tags:
        - Library
  /library:
    post:
      summary: Library creation
      operationId: addLibrary
      responses:
        '201':
          description: Library profile created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Library'
        '400':
          description: Bad Request - missing or invalid input values provided
        '500':
          description: Error creating the library profile
      description: Add an additional partner library to facilitate future inter-library loans
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LibraryInfo'
        description: Provide basic details for creating a library profile
      tags:
        - Library
    get:
      summary: Retrieve this library's info
      operationId: getThisLibrary
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Library'
        '500':
          description: An error occured while retrieving this library's details
      description: Retrieve this library's details
      tags:
        - Library
  /libraries:
    get:
      summary: Retrieve all library details
      tags:
        - Library
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Library'
        '500':
          description: An error occurred while retrieving all library details
      operationId: getAllLibraries
      description: Retrieve information about all libraries
  /rentals:
    get:
      summary: Retrieve all current rentals
      tags:
        - Rental
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Loan'
        '500':
          description: Error occurred while retrieving all rental records
      operationId: getAllRentals
      description: Retrieve all current rentals
      parameters:
        - schema:
            type: boolean
            default: 'false'
          in: query
          name: includeReturned
          description: Specifies whether to include records of already returned books as well
          allowEmptyValue: true
components:
  schemas:
    UserInfo:
      title: User Info
      type: object
      description: ''
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        isStaff:
          type: boolean
          description: Specifies whether the user is a library staff
        subscriptionDate:
          type: string
          format: date
          description: Joining date as milliseconds from the epoch or in ISO 8601 timestamp format
        gender:
          type: string
          enum:
            - MALE
            - FEMALE
            - TRANSGENDER
            - NEUTRAL
            - NON_BINARY
            - AGENDER
            - PANGENDER
            - GENDERQUEER
            - TWO_SPIRIT
            - THIRD_GENDER
            - ALL
            - NONE
        isLibrarian:
          type: boolean
          description: Defines whether this user is a librarian staff member
      required:
        - firstName
        - lastName
        - email
        - gender
    User:
      title: User
      type: object
      description: ''
      x-examples:
        David Osideko:
          id: 1
          firstName: David
          lastName: Osideko
          email: david.osideko@hsbc.com
          isStaff: true
          subscriptionDate: 1616936513599
          gender: MALE
          isLibrarian: true
      properties:
        id:
          type: integer
          description: Unique identifier for the user
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        isStaff:
          type: boolean
          description: Specifies whether the user is a library staff
        subscriptionDate:
          type: string
          format: date
          description: Joining date as milliseconds from the epoch or in ISO 8601 timestamp format
        gender:
          type: string
          enum:
            - MALE
            - FEMALE
            - TRANSGENDER
            - NEUTRAL
            - NON_BINARY
            - AGENDER
            - PANGENDER
            - GENDERQUEER
            - TWO_SPIRIT
            - THIRD_GENDER
            - ALL
            - NONE
        isLibrarian:
          type: boolean
          description: Defines whether this user is a librarian staff member
      required:
        - firstName
        - lastName
        - email
...
    LoanInfo:
      title: Loan Info
      type: object
      properties:
        bookUuid:
          type: string
          description: Unique identifier for the book on loan
        fromDate:
          type: string
          description: Date the book was borrowed
          format: date
        dueDate:
          type: string
          description: Date the book is expected to be returned
          format: date
        returnDate:
          type: string
          description: Actual return date of the book
          format: date
        chargeDue:
          type: number
          description: Charges due to late returns or loss of the book
          format: double
          default: 0
          minimum: 0
        chargePaid:
          type: number
          description: Amount paid by the borrower
          default: 0
          format: double
        memberId:
          type: integer
          description: 'User id, if the book was borrowed by a member'
        libraryId:
          type: integer
          description: 'Library id, if this is an inter-library loan'
      required:
        - bookUuid
    Loan:
      title: Loan
      type: object
      x-examples:
        Loan to David:
          id: 1
          bookUuid: c5043628-d7d5-4969-b6b9-9ad8bb2a57b7
          fromDate: 1616954323
          dueDate: 1619619584
          returnDate: 0
          chargeDue: 0
          chargePaid: 0
          memberId: 10
          libraryId: 0
      properties:
        id:
          type: integer
        bookUuid:
          type: string
          description: Unique identifier for the book on loan
        fromDate:
          type: string
          description: Date the book was borrowed
          format: date
        dueDate:
          type: string
          description: Date the book is expected to be returned
          format: date
        returnDate:
          type: string
          description: Actual return date of the book
          format: date
        chargeDue:
          type: number
          description: Charges due to late returns or loss of the book
          format: double
          default: 0
          minimum: 0
        chargePaid:
          type: number
          description: Amount paid by the borrower
          default: 0
          format: double
        memberId:
          type: integer
          description: 'User id, if the book was borrowed by a member'
        libraryId:
          type: integer
          description: 'Library id, if this is an inter-library loan'
      required:
        - bookUuid
    LibraryInfo:
      title: Library Info
      type: object
      x-examples: {}
      properties:
        name:
          type: string
          description: Library name
        address:
          type: string
          description: Full address of the library
        email:
          type: string
          format: email
          description: Contact email
        phoneNumber:
          type: string
          description: Contact phone number
    Library:
      title: Library
      type: object
      x-examples:
        HSBC Staff Library:
          id: 1
          name: HSBC Staff Library
          address: Somewhere in the UK
          email: staff.library@hsbc.co.uk
          phoneNumber: 1227123456
          isSelf: true
      properties:
        id:
          type: integer
        name:
          type: string
          description: Library name
        address:
          type: string
          description: Full address of the library
        email:
          type: string
          format: email
          description: Contact email
        phoneNumber:
          type: string
          description: Contact phone number
        self:
          type: boolean
          description: Specifies whether the object details refer to this service owner i.e. HSBC
    BookInfo:
      title: Book Info
      description: ''
      type: object
      x-examples:
        Power of Now:
          author: Eckhart Tolle
          title: The Power of Now
          description: A Guide to Spiritual Enlightenment
          isbn: '9780340733509'
          publicationDate: 0
          language: ENGLISH
          coverImageData: ''
          genre: SELF_HELP
          classification: B
          copiesAvailable: 1
      properties:
        author:
          type: string
        description:
          type: string
        isbn:
          type: string
          description: International Standard Book Number
        publicationDate:
          type: string
          description: Publication date in milliseconds from the epoch or in ISO 8601 timestamp format
        coverImageData:
          type: string
          description: Base64 encoded cover image
        language:
          type: string
          enum:
            - ENGLISH
            - FRENCH
        genre:
          type: string
          description: One or more genres relating to this book
          enum:
            - FICTION
            - NON_FICTION
            - SELF_HELP
            - CLASSIC
            - REFERENCE
            - JOURNAL
            - MAGAZINE
        classification:
          type: string
          description: The Library of Congress Classification
          enum:
            - A
            - B
            - C
            - D
            - E
            - F
            - G
            - H
            - J
            - K
            - L
            - M
            - 'N'
            - P
            - Q
            - R
            - S
            - T
            - U
            - V
            - Z
        copiesAvailable:
          type: integer
          description: The number of available copies of this book
        title:
          type: string
      required:
        - author
        - isbn
        - language
        - genre
        - classification
        - copiesAvailable
        - title
...