# David Osideko- HSBC/src/main/java/uk/co/hsbc/library/service/model/LibraryInfo.java
package uk.co.hsbc.library.service.model;

import lombok.Getter;
import lombok.Setter;

@Getter
@Setter
public class LibraryInfo {

	private String name;
	private String address;
	private String email;
	private String phoneNumber;
}